RavenDB - 17262 - last changes
