+++ src/Raven.Server/Commercial/LicenseHelper.cs
@@ -255,7 +255,7 @@ namespace Raven.Server.Commercial
 
                 var modifiedJsonObj = context.ReadObject(settingsJson, "modified-settings-json");
 
-                var indentedJson = JsonStringHelper.IndentJsonString(modifiedJsonObj.ToString());
+                var indentedJson = JsonStringHelper.Indent(modifiedJsonObj.ToString());
                 SettingsZipFileHelper.WriteSettingsJsonLocally(_serverStore.Configuration.ConfigPath, indentedJson);
 
                 return true;
