+++ src/Raven.Server/Commercial/SetupWizard/LetsEncryptSetupUtils.cs
@@ -4,20 +4,21 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Raven.Client.Documents.Operations;
+using Raven.Server.Commercial;
+using Raven.Server.Commercial.LetsEncrypt;
 
-namespace Raven.Server.Commercial.LetsEncrypt
+namespace rvn.Server.SetupWizard;
+
+public class LetsEncryptSetupUtils
 {
-    public static class LetsEncryptUtils
-    {
         private const string AcmeClientUrl = "https://acme-v02.api.letsencrypt.org/directory";
 
-        public static async Task<byte[]> SetupLetsEncrypt(SetupInfo setupInfo,  SetupProgressAndResult progress, CancellationToken token)
+        public static async Task<byte[]> Setup(SetupInfo setupInfo,  SetupProgressAndResult progress, CancellationToken token)
         {
             progress.Processed++;
             progress?.AddInfo("Setting up RavenDB in Let's Encrypt security mode.");
 
-            if (EmailValidator.IsValidEmail(setupInfo.Email) == false)
+            if (EmailValidator.IsValid(setupInfo.Email) == false)
                 throw new ArgumentException("Invalid e-mail format" + setupInfo.Email);
 
             var acmeClient = new LetsEncryptClient(AcmeClientUrl);
@@ -124,49 +125,6 @@ namespace Raven.Server.Commercial.LetsEncrypt
             }
         }
 
-        public static async Task<byte[]> ImportCertificateSetup(SetupInfo setupInfo, SetupProgressAndResult progress, CancellationToken token)
-        {
-            try
-            {
-                progress.Processed++;
-                progress?.AddInfo("Setting up RavenDB in 'Secured Mode'.");
-                progress?.AddInfo("Starting validation.");
-
-                if (EmailValidator.IsValidEmail(setupInfo.Email) == false)
-                    throw new ArgumentException("Invalid e-mail format: " + setupInfo.Email);
-
-                byte[] zipFile;
-                try
-                {
-                     zipFile = await SettingsZipFileHelper.CompleteClusterConfigurationAndGetSettingsZip(new CompleteClusterConfigurationParameters
-                    {
-                        Progress = progress,
-                        SetupInfo = setupInfo,
-                        SetupMode = SetupMode.None,
-                        LicenseType = LicenseType.None,
-                        Token = CancellationToken.None,
-                        CertificateValidationKeyUsages = true
-                    });
-                     
-                     progress.Processed++;
-                     progress.AddInfo("Configuration settings created.");
-                     progress.AddInfo("Setting up RavenDB in 'Secured Mode' finished successfully.");
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException("Failed to create the configuration settings.", e);
-                }
-
-                return zipFile;
-            }
-            catch (Exception e)
-            {
-                const string str = "Setting up RavenDB in 'Secured Mode' failed.";
-                progress?.AddError(str, e);
-                throw new InvalidOperationException(str, e);
-            }
-        }
-
         public static async Task<(string Challenge, LetsEncryptClient.CachedCertificateResult Cache)> InitialLetsEncryptChallenge(
             SetupInfo setupInfo,
             LetsEncryptClient client,
@@ -190,5 +148,4 @@ namespace Raven.Server.Commercial.LetsEncrypt
                 throw new InvalidOperationException("Failed to receive challenge(s) information from Let's Encrypt.", e);
             }
         }
-    }
 }
