+++ src/Raven.Server/Web/System/SetupHandler.cs
@@ -472,7 +472,7 @@ namespace Raven.Server.Web.System
                     writer.WriteStartArray();
 
                     var first = true;
-                    foreach (var value in LetsEncryptCertificateUtil.GetCertificateAlternativeNames(certificate))
+                    foreach (var value in CertificateUtils.GetCertificateAlternativeNames(certificate))
                     {
                         if (first == false)
                             writer.WriteComma();
@@ -555,7 +555,7 @@ namespace Raven.Server.Web.System
 
                 var modifiedJsonObj = context.ReadObject(settingsJson, "modified-settings-json");
 
-                var indentedJson = JsonStringHelper.IndentJsonString(modifiedJsonObj.ToString());
+                var indentedJson = JsonStringHelper.Indent(modifiedJsonObj.ToString());
                 SettingsZipFileHelper.WriteSettingsJsonLocally(ServerStore.Configuration.ConfigPath, indentedJson);
             }
